atos-deps: fixed issue with multiple archive at link time.
This changes fixes two problem in atos-deps.
First, the multiple occurence of a library archive on the
command line must be allowed. This change allows multiple
occurence of the same input in the dependence graph generation.
Second, when in presence of shared lib, the shared lib version must
be checked before the archive version. This change implements this
search. Note though that shared libs are still ignored as input,
this requires a separate development.