utils.py: factorized wrapper script creation.
Added a generate_script() function to atos_lib that creates
a script that wraps a command, including environment and
chdir to the current working directory.
Added a sh_quote kwarg argument to process.list2cmdline()
for generating quoted arguments.
Quoted arguments are more verbose than the default backslashed
scheme but are more robust to arbitrary char sequences as
found in environment.