atos-cookie: added script for generating cookies.
The atos-cookie script outputs a unique cookie to stdout.
It can be used by the user to generate a cookie for an
exploration, the cookie is then passed with the --cookie
option.
If some preexisting cookies are given as argument a
digest of the cookies is generated.
Added unique_cookie() function in atos lib and factorized
code with generators.py.