atos-run: fix size computation for non absolute exec path.
When the executable path was not absolute, the executable was
search only in PATH.
Actually if the relative path exists, use it directly.
For instance in force mode:
atos-init -f ...
atos-deps my_exe
This will add my_exe to the targets list and if my_exe is in current dir
atos-run will succeed at finding it for running size.