test070.py: added expected to fail test.
This test exhibits an issue with the file by file mapping from functions to
objects.
This mapping is not sufficient for discriminating executables/libs in the
profiling output.
If several executable contain a main function that is hot for only one of
them, all objects will be marked as hot, while it seems feasible to
mark only the object for the actually hot executable.