atos-play: added --options and --variant flags.
Added -w/--variant flag to atos-play for playing a variant id.
Added -a/--options flag to atos-play for playing a variant
while adding additional flags to the build. This is useful
for building a variant with some additional flags for refining
a result manually.
Fixed reference variant for speedups in atos_lib.get_results(),
use opts.refid instead of 'REF'.
Improved coverage of atos-play in test033.sh.