atoslib: rewrite storage and computation of dependency graph.
This is a rewrite of the dependency graph computation that obsoletes
the previous atos_deps module.
The previous behavior is reproducible with the --legacy flag passed
to atos-init and other tools.
Each compilation command executed by the driver is now recorded
in an object database as a recipe object where are stored all
input files and their hash values, all the commands environment,
and all ouput files and their hash value.
When generating the build makefile with atos-deps the graph is
reconstructed from this object storage.
The generated makefile is now working with recipes digest values
instead of actual pathnames, this removes the issues related to
multiple occurences of the same paths.
This is a basis for completing the support for duplicate objects
construction and moves between build phases.
The current object storage implementation should be latter
optimized with a cache of already loaded objects as for now the
atos-deps process time has significantly increased.