globals.py: fix default regexps for cc drivers and archivers.
Be more strict when detecting compiler drivers and archivers through
regexps.
The previous regexp match for instance include .*cc.
This implied that rcc for instance was considered as a compiler
driver which is not the case, or tar was considered as a bare
archiver which is also wrong.
Now the scheme is to detect predetermined base names or specific
cross target prefixes followed by these basenames or a generic prefix
followed by the basenames, actually '.*-'.

Fixes: http://hostname/tracker/?func=detail&aid=207104