atos_graph.py: reimplemented atos-graph --follow with signals.
Implements a timer based on SIGALARM in atos_lib.repeaetalarm class.
Uses this class for atos-graph --follow implementation.
In Tk matplotlib backend, calling draw() from another thread
of execution does not work on all platforms, as per the previous
implementation with the thread based repeattimer class.
Also the native mathplotlib support for timers work but it is is not
available in old matplotlib packages.
The signal based approach seems to work correctly and the
restriction (only one alarm) is not an issue for atos-graph.

Fixes: https://codex.cro.st.com/tracker/?func=detail&aid=208156