atos-deps: allow selection of targets to be built.
The atos-configurations/targets file can be modified to remove
or comment out (with '#') some targets determined by
atos-init or atos-deps --all.
In this case atos-deps must be run again with no arguments to
reconstruct the build informations.
This allow to remove useless targets, or targets that should not
be optimized or targets that fail to rebuild with atos-build.

Implements: http://hostname/tracker/?func=detail&aid=207974