progress.py: fixed output of remaining time.
Remaining time at very begining and very end of build/run
was displayed as ??:??:??.
Fixed in the _display_progress() function, actually at the very
begining the remaining time is the estimated time and at the
very end it is 0.

Also padded some field to have a more stable output (percentages
and progress_type).

Fixes: http://hostname/tracker/?func=detail&aid=209825