progress.py: ensure that progress lines are correctly cleared.
The message() function in logger.py when detecting a tty will
always clear the full line before outputing the message string.
This ensure that when interleaving progress messages with normal
messages there is no trailing progress message left at the end of
a message line.
Also the handle_interrupt() function in __init__.py now force a
line return on stdout and flushes stdout and stderr.

Fixes: http://codex.cro.st.com/tracker/?func=detail&aid=209839