atos-driver.py: skip blacklisted recipes and populate blacklist.
In the case where a build recipes fails, a recipes blacklist is
populated when --blacklist is passed to atos-build.
This is the default in atos-init when building the reference variant.
The blacklist is inpected by atos-driver in order to skip recipe
commands that are blacklisted.
This solves all build system issues related to missing files,
though it may lead to explorations where objects are actually not
optimized.

Fixes: http://hostname/tracker/?func=detail&aid=207858