atos-proot: added proot wrapper script.
The script atos-proot is a wrapper fo proot that can be used
by the user directly in place of proot itself.
ATOS calls atos-proot instead of proot directly.
The atos-proot script will invoke $ATOS_PROOT if given
or the proot in the ATOS installation, or lastly the proot in
the path.
If non empty, $ATOS_PROOT_ARGS will be passed before the
user specified command line arguments.