obj_storage: fixed issue with storing of blobs.
The storage of blob was done through hard link when this was
possible to hard link a source file into the build.stg/blobs
directory.
This optimization is wrong as an input file such as a .a can
be modified in place, and thus the actual content of the stored
copy in the atos storage is also modified in case of hard link.

For example a ranlib libfoo.a after an ar rc libfoo.a will
generate an inconsistent blob for libfoo.a which will be
stored with an expected sha1sum after the ar, but then
modified in place in the storage by the ranlib.

The hard link optimization was used in two places:
obj_storage.py for the storage of the input/output blobs, and
recipes.py for the retrieval of the input ot output files from
the object storage. Both have been removed.

Refs: http://codex.cro.st.com/tracker/?func=detail&aid=218474
Fixes: http://codex.cro.st.com/tracker/?func=detail&aid=219386