recipes.py: create dir before fetching inputs/outputs
The input files backed up in the ATOS object storage can
be used when the actual input file has disappeared from the build.
Also the output files backep up into the ATOS storage can be
used when the recipe is blacklisted as a fallback.
In both cases, the actual directory of the input/output file may also
have been cleaned up by the build system.
Thus, when fetching from the object storage inputs/outputs files the
hosting directory must be created first.

This fixes a bug in the libparrot build.

Fixes: http://codex.cro.st.com/tracker/?func=detail&aid=219486