plugins: added support for compilation of plugins in atos-init.
Now, the plugins are tentatively compiled while configuring
the compiler in atos-init or atos-config.
If the compiled plugin is correctly working it is used in place
of prebuilt plugins.
Augmented the minimal test for plugin such that the offset of
the global parameters flags is checked. This detects earlier
that the plugin will not support param flags and will avoid
latter segfaults when using the plugin in atos-explore-acf.

Fixes: http://codex.cro.st.com/tracker/?func=detail&aid=219823