plugins: reworked plugins build to use the gcc-plugins repo
This change remove the plugin sources reference from the repository
and adds a dependency on the now externalized gcc-plugins repo.

The plugins sources are extracted from the gcc-plugins repository
(ref to DEPENDENCIES files) and copied locally first.
Thus the plugin sources are locally updated but are a copy of the
external gcc-plugins repo.
The sources are copied into plugins/acf-plugin/plugins-src and
should not be modified there.

The steps for updating plugins has not changed, though, now the
'make prep' phase does a copy of the gcc-plugins sources first:
$ cd plugins/acf-plugins
$ ./dependencies extract
$ make prep

In order to update the sources from the gcc-plugins repo, do:
$ cd plugins/acf-plugins
$ ./dependencies update
xafter optionally modifying the DEPENDENCIES file to point to a new
gcc-plugins version.

The support for gcc-4.5.2 has been dropped as it is not maintained
anymore in gcc-plugins.

The build system still builds the plugins for gcc-4.6.2 and gcc-4.7.2
from the preprocessed files and install these default plugins.

The atos-init tool will still try to compile the acf-plugin with
the actual build compiler and fall back to the default plugin if it fails.

Note that the default plugins do not always work on some cross compilers,
thus in the future it may be an option to completly drop the support for
prebuilt plugins and rely only on the user's compiler capacity to build
plugins.

This new build scheme fixes some issues still present in the old plugin
build system. In particular is fixes an issue with the acf-plugion build
on armv7 cross compilers.

Fixes: http://hostname/tracker/?func=detail&aid=236189