atos-config: imply --print-cfg when --compiler is specified
Changed atos-config such that --print-cfg is implied (instead of
updating the atos-configurations/config.json file) when
--compiler option is passed.
This avoids issues with missing configurations files and
remove side effects when atos-config is used on a specific compiler
independently of any previous atos-audit.

Modified tests for this new behavior and for completing code coverage.