atos-driver: pass optimization to all final link commands
This change reverts a behavior introduced in revision 63b65d31.
With this former change, the optimization flags were passed only
for configured targets (whatever the kind: exec, shared libs or
reloc links).

Reverted to pass optimization to all targets as before.
The only exception for now is that for relocatable links, the default is
to not pass optimization flags unless it is a configured target
(either by the user -e option or because it is determined to be a final
relocatable link).

Added a test that compilation does not fail in the case where an object is
shared between two executables, and only one of the two is configured.

Fixes: http://codex.cro.st.com/tracker/?func=detail&aid=242742